rename custom `stablehlo` folder to `hlo-extensions`
