+++ mlir/include/CMakeLists.txt
@@ -1,9 +1,9 @@
 add_subdirectory(Catalyst)
 add_subdirectory(Gradient)
+add_subdirectory(hlo-extensions)
 add_subdirectory(Ion)
 add_subdirectory(MBQC)
 add_subdirectory(Mitigation)
 add_subdirectory(QEC)
 add_subdirectory(Quantum)
-add_subdirectory(stablehlo)
 add_subdirectory(Test)
