+++ mlir/lib/CMakeLists.txt
@@ -2,10 +2,10 @@ add_subdirectory(CAPI)
 add_subdirectory(Catalyst)
 add_subdirectory(Driver)
 add_subdirectory(Gradient)
+add_subdirectory(hlo-extensions)
 add_subdirectory(Ion)
 add_subdirectory(MBQC)
 add_subdirectory(Mitigation)
 add_subdirectory(QEC)
 add_subdirectory(Quantum)
-add_subdirectory(stablehlo)
 add_subdirectory(Test)
