+++ mlir/lib/hlo-extensions/hlo_custom_call_lowering.cpp
@@ -17,18 +17,16 @@
 #include <vector>
 
 #include "llvm/Support/Debug.h"
-
-#include "stablehlo/dialect/StablehloOps.h"
-#include "stablehlo/transforms/Passes.h"
-
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/Index/IR/IndexDialect.h"
 #include "mlir/Pass/Pass.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
+#include "stablehlo/dialect/StablehloOps.h"
+#include "stablehlo/transforms/Passes.h"
 
 #include "Catalyst/IR/CatalystDialect.h"
-#include "stablehlo/Passes.h"
-#include "stablehlo/Patterns.h"
+#include "hlo-extensions/Passes.h"
+#include "hlo-extensions/Patterns.h"
 
 using namespace llvm;
 using namespace mlir;
@@ -36,7 +34,7 @@ using namespace catalyst;
 
 namespace catalyst {
 #define GEN_PASS_DEF_HLOCUSTOMCALLLOWERINGPASS
-#include "stablehlo/Passes.h.inc"
+#include "hlo-extensions/Passes.h.inc"
 
 struct HloCustomCallLoweringPass : impl::HloCustomCallLoweringPassBase<HloCustomCallLoweringPass> {
     using HloCustomCallLoweringPassBase::HloCustomCallLoweringPassBase;
