+++ mlir/lib/hlo-extensions/scatter_lowering.cpp
@@ -17,19 +17,17 @@
 #include <vector>
 
 #include "llvm/Support/Debug.h"
-
-#include "stablehlo/dialect/StablehloOps.h"
-#include "stablehlo/transforms/Passes.h"
-
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/Index/IR/IndexDialect.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
 #include "mlir/Dialect/Tensor/IR/Tensor.h"
 #include "mlir/Pass/Pass.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
+#include "stablehlo/dialect/StablehloOps.h"
+#include "stablehlo/transforms/Passes.h"
 
-#include "stablehlo/Passes.h"
-#include "stablehlo/Patterns.h"
+#include "hlo-extensions/Passes.h"
+#include "hlo-extensions/Patterns.h"
 
 using namespace llvm;
 using namespace mlir;
@@ -37,7 +35,7 @@ using namespace catalyst;
 
 namespace catalyst {
 #define GEN_PASS_DEF_SCATTERLOWERINGPASS
-#include "stablehlo/Passes.h.inc"
+#include "hlo-extensions/Passes.h.inc"
 
 struct ScatterLoweringPass : impl::ScatterLoweringPassBase<ScatterLoweringPass> {
     using ScatterLoweringPassBase::ScatterLoweringPassBase;
