+++ mlir/lib/hlo-extensions/stablehlo_legalize_sort.cpp
@@ -41,6 +41,7 @@ limitations under the License.
 #include <memory>
 #include <utility>
 
+#include "llvm/ADT/STLExtras.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/Arith/Utils/Utils.h"
 #include "mlir/Dialect/Bufferization/IR/Bufferization.h"
@@ -63,9 +64,8 @@ limitations under the License.
 #include "mlir/Transforms/DialectConversion.h"
 #include "stablehlo/dialect/StablehloOps.h"
 #include "stablehlo/transforms/Passes.h"
-#include "llvm/ADT/STLExtras.h"
 
-#include "stablehlo/Passes.h"
+#include "hlo-extensions/Passes.h"
 
 using namespace mlir;
 using namespace stablehlo;
@@ -75,7 +75,7 @@ namespace catalyst {
 
 #define GEN_PASS_DEF_STABLEHLOLEGALIZESORTPASS
 #define GEN_PASS_DECL_STABLEHLOLEGALIZESORTPASS
-#include "stablehlo/Passes.h.inc"
+#include "hlo-extensions/Passes.h.inc"
 
 } // namespace catalyst
